+++ packages/pixels/src/Drawer/index.ts
@@ -1,0 +1,25 @@
+import Drawer, {
+  drawerBackdrops,
+  drawerPlacements,
+  drawerRadii,
+  drawerSizes,
+  drawerVariants,
+} from './Drawer';
+
+export type {
+  DrawerProps,
+  DrawerSizes,
+  DrawerRadius,
+  DrawerBackdrop,
+  DrawerPlacement,
+} from './Drawer';
+
+export {
+  drawerBackdrops,
+  drawerPlacements,
+  drawerRadii,
+  drawerSizes,
+  drawerVariants,
+};
+
+export default Drawer;